// libs/alert-core/src/lib/constants/alert-activity-type.ts
export enum AlertActivityType {
  AlertCreated = 'ALERT_CREATED',
  AlertGrouped = 'ALERT_GROUPED',
  AlertAssigned = 'ALERT_ASSIGNED',
  AlertReassigned = 'ALERT_REASSIGNED',
  AlertStateChanged = 'ALERT_STATE_CHANGED',
  EvidenceAddedOrUpdated = 'EVIDENCE_ADDED_OR_UPDATED',
  AlertResolved = 'ALERT_RESOLVED',
  AlertDismissed = 'ALERT_DISMISSED',
  SLABreached = 'SLA_BREACHED',
  NoteAdded = 'NOTE_ADDED',
  PriorityChanged = 'PRIORITY_CHANGED',
}